+++ server/impl/src/main/java/com/walmartlabs/concord/server/agent/AgentCommandWatchdog.java
@@ -21,7 +21,7 @@ package com.walmartlabs.concord.server.agent;
  */
 
 import com.walmartlabs.concord.db.AbstractDao;
-import com.walmartlabs.concord.server.BackgroundTask;
+import com.walmartlabs.concord.server.PeriodicTask;
 import com.walmartlabs.concord.server.agent.AgentCommand.Status;
 import org.jooq.Configuration;
 import org.jooq.Field;
@@ -39,7 +39,7 @@ import static org.jooq.impl.DSL.field;
 
 @Named
 @Singleton
-public class AgentCommandWatchdog implements BackgroundTask {
+public class AgentCommandWatchdog extends PeriodicTask {
 
     private static final Logger log = LoggerFactory.getLogger(AgentCommandWatchdog.class);
 
@@ -48,61 +48,17 @@ public class AgentCommandWatchdog implements BackgroundTask {
 
     private final WatchdogDao watchdogDao;
 
-    private Thread worker;
-
     @Inject
     public AgentCommandWatchdog(WatchdogDao watchdogDao) {
+        super(POLL_DELAY, ERROR_DELAY);
         this.watchdogDao = watchdogDao;
     }
 
     @Override
-    public void start() {
-        this.worker = new Thread(new Worker(POLL_DELAY, ERROR_DELAY, watchdogDao),
-                "stalled-agent-commands-worker");
-        this.worker.start();
-    }
-
-    @Override
-    public void stop() {
-        if (this.worker != null) {
-            this.worker.interrupt();
-        }
-    }
-
-    private static final class Worker implements Runnable {
-
-        private final long interval;
-        private final long errorInterval;
-        private final WatchdogDao watchdogDao;
-
-        private Worker(long interval, long errorInterval, WatchdogDao watchdogDao) {
-            this.interval = interval;
-            this.errorInterval = errorInterval;
-            this.watchdogDao = watchdogDao;
-        }
-
-        @Override
-        public void run() {
-            while (!Thread.currentThread().isInterrupted()) {
-                try {
-                    Field<Timestamp> cutoff = currentTimestamp().minus(field("interval '10 minutes'"));
-                    int n = watchdogDao.failStalled(cutoff);
-                    log.info("run -> {} command(s) are timed out", n);
-                    sleep(interval);
-                } catch (Exception e) {
-                    log.error("run -> error: {}", e.getMessage(), e);
-                    sleep(errorInterval);
-                }
-            }
-        }
-
-        private void sleep(long ms) {
-            try {
-                Thread.sleep(ms);
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            }
-        }
+    protected void performTask() {
+        Field<Timestamp> cutoff = currentTimestamp().minus(field("interval '10 minutes'"));
+        int n = watchdogDao.failStalled(cutoff);
+        log.info("run -> {} command(s) are timed out", n);
     }
 
     @Named