+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- concord-server: periodic audit log clean up;
+- concord-server: option to disable audit logging;
 - concord-server: added limits for "encrypted strings" size;
 - concord-server: improved error reporting for non-process related
 errors;